feat: added a verification in delete user to know if user exists or not

+++ src/main/java/com/academico/webproject/service/UserService.java
@@ -22,7 +22,11 @@
         }
 
         public void deleteUser(String id) {
-            userRepository.deleteById(id);
+            if (userRepository.existsById(id)) {
+                userRepository.deleteById(id);
+            } else {
+                throw new RuntimeException("Band not found with ID: " + id);
+            }
         }
 
         public User updateUser(String id, User userDetails) {
